+++ backend_go/surface_intersect/go.mod
@@ -1,4 +1,4 @@
-module intersectTest
+module surface_intersect
 
 go 1.20
