+++ backend_go/surface_intersect/xtgeo/calculate_vector_length_and_angle.go
