+++ backend_go/surface_intersect/xtgeo/constants.go
