+++ backend_go/surface_intersect/xtgeo/distance_between_point_and_line.go
