+++ backend_go/surface_intersect/xtgeo/projected_point_on_line.go
