+++ backend_go/surface_intersect/xtgeo/surface_ijk_to_1d.go
